cpuid max leaf should return 32-bits, not 8 (esp. for AMD)

+++ include/nautilus/cpuid.h
@@ -26,7 +26,7 @@
 #include <nautilus/intrinsics.h>
 #include <nautilus/naut_types.h>
 
-uint8_t cpuid_leaf_max(void);
+uint32_t cpuid_leaf_max(void);
 uint8_t cpuid_get_family(void);
 uint8_t cpuid_get_model(void);
 uint8_t cpuid_get_step(void);
